+++ src/components/folder/FolderMaker/FolderMaker.tsx
@@ -3,7 +3,7 @@
   만약 화면의 크기가 tablet 미만이 되면 floating button 형태로 렌더링됨.
 */
 import FolderAddModal from "@/modals/FolderAddModal/FolderAddModal";
-import { useModalStore } from "@/store/useModalStore";
+import { useModalStore } from "@/store/ModalStore";
 
 import styles from "./FolderMaker.module.scss";
