+++ src/modals/FolderDeleteModal/FolderDeleteModal.tsx
@@ -4,9 +4,9 @@ import { useQueryClient, useMutation } from "@tanstack/react-query";
 import { deleteFolder } from "@/api/getFolderCRUDApi";
 import useToast from "@/hooks/useToast";
 import ModalCreator from "@/modals/ModalCreator";
+import { useModalStore } from "@/store/ModalStore";
 
 import styles from "./FolderDeleteModal.module.scss";
-import { useModalStore } from "@/store/useModalStore";
 
 interface FolderDeleteModalProps {
   folderId: string;
